+++ pambox/speech/mrsepsm.py
@@ -41,6 +41,7 @@ class MrSepsm(Sepsm):
 
     """
 
+    # Default center frequencies of the modulation filterbank.
     _default_modf = (1., 2., 4., 8., 16., 32., 64., 128., 256.)
 
     def __init__(self, fs=22050, cf=Sepsm._default_center_cf,
@@ -58,23 +59,29 @@ class MrSepsm(Sepsm):
         self.name = name
         self.snr_env_ceil = snr_env_ceil
 
-    def _mr_env_powers(self, channel_env, filtered_envs):
-        """Calcultes the envelope power in multi-resolution windows.
+    def _mr_env_powers(self, channel_envs, filtered_envs):
+        """Calculates the envelope power in multi-resolution windows.
 
         Parameters
         ----------
         channel_env : ndarray
-            Envelope of the peripheral channel.
+            Envelope of the peripheral channel. The shape should be (N_SIG,
+            N_CHAN, N)
         filtered_envs : ndarray
-            Filtered envelope.
+            Filtered envelope. The shape should be (N_SIG, N_CHAN, N_MODF, N)
 
         Returns
         -------
-        masked_array
-            Multi-resolution envelope powers of shape N_mod_channels x Length.
+        mr_env_powers : masked_array
+            Multi-resolution envelope powers of shape (N_SIG, N_CHAN, N_MODF,
+            N_SEG), where N_SEG is the maximum number of segments in the
+            multi-resolution process, which is for the highest modulation
+            center frequency. For a 1 sec sample, it is about 400 segments.
             Low modulation frequencies come first.
 
         """
+        # Here we find the duration and the number of windows for each
+        # modulation center frequency...
         len_env = filtered_envs.shape[-1]
         win_durations = 1. / np.asarray(self.modf, dtype='float')
         if self.min_win is not None:
@@ -83,40 +90,56 @@ class MrSepsm(Sepsm):
                                .downsamp_factor).astype('int')
         n_segments = np.ceil(len_env / win_lengths).astype('int')
 
-        # DC power used for normalization. Divided by 2 such that a fully
-        # modulated signal has # an AC-power of 1.
-        dc_power = np.mean(channel_env) ** 2 / 2
+        # ... then we calculate the DC power used for the normalization.
+        # We divide it by 2 such that a fully modulated signal has an
+        # AC-power of 1...
+        dc_power = np.mean(channel_envs, axis=-1) ** 2 / 2
 
-        # Create a masked array of zeros, where all entries are hidden.
-        mr_env_powers = np.ma.masked_all((len(self.modf), max(n_segments)))
+        # ... then we create a masked array of zeros, where all entries are
+        # hidden...
+        mr_env_powers = np.ma.masked_all(
+            (filtered_envs.shape[:-1] + (np.max(n_segments),))
+        )
 
-        for i_modf, (n_seg, env, win_length) in enumerate(
-                zip(n_segments, filtered_envs, win_lengths)):
-            n_complete_seg = n_seg - 1
-            last_idx = int(n_complete_seg * win_length)
-            # Reshape to n_seg x win_length so that we can calculate the
-            # variance in a single operation
-            tmp_env = env[:last_idx].reshape((-1, win_length))
-            # Normalize the variance by N-1, like in MATLAB.
-            tmp_env_powers = np.var(tmp_env, axis=-1, ddof=1) / dc_power
-            # Treat the last segment independently, just in case it is not
-            # complete, i.e. it is shorter than the window length.
-            tmp_env_powers_last = np.var(env[last_idx:], ddof=1) / dc_power
-            mr_env_powers[i_modf, :n_complete_seg] = tmp_env_powers
-            mr_env_powers[i_modf, n_complete_seg] = tmp_env_powers_last
-
-            mr_env_powers.mask[i_modf, :n_seg] = False
+        # ... and start looping throug the input signals...
+        for i_sig, sig_envs in enumerate(filtered_envs):
+            # ... and the channels...
+            for i_chan, chan_envs in enumerate(sig_envs):
+                # ... and the modulation channels, zipped with the window
+                # lengths and number of segments...
+                for i_modf, (n_seg, win_length, env) in enumerate(
+                        zip(n_segments, win_lengths, chan_envs)):
+                    n_complete_seg = n_seg - 1
+                    last_idx = int(n_complete_seg * win_length)
+                    # ... we reshape to n_seg x win_length so that we can
+                    # calculate the variance in a single operation...
+                    tmp_env = env[:last_idx].reshape((-1, win_length))
+                    # ... then we normalize the variance by N-1, like in MATLAB.
+                    tmp_env_powers = np.var(tmp_env, axis=-1, ddof=1) / \
+                                     dc_power[i_sig, i_chan]
+                    # ... and treat the last segment independently, just in
+                    # case it is not # complete, i.e. it is shorter than the
+                    # window length.
+                    tmp_env_powers_last = np.var(env[last_idx:], ddof=1) / \
+                                          dc_power[i_sig, i_chan]
+                    # ... then we save it in our masked_array...
+                    mr_env_powers[i_sig, i_chan, i_modf, :n_complete_seg] = \
+                        tmp_env_powers
+                    mr_env_powers[i_sig, i_chan, i_modf, n_complete_seg] = \
+                        tmp_env_powers_last
+                    # ... and finally make these values visible through the
+                    # mask.
+                    mr_env_powers.mask[i_sig, i_chan, i_modf, :n_seg] = False
 
         return mr_env_powers
 
     @staticmethod
     def _time_average(mr_snr_env):
-        """
+        """Averages the multi-resolution SNRenv over time.
 
         Parameters
         ----------
         mr_snr_env : ndarray
-
 
         Returns
         -------
@@ -126,7 +149,7 @@ class MrSepsm(Sepsm):
         return mr_snr_env.mean(axis=-1)
 
     def _mr_snr_env(self, p_mix, p_noise):
-        """Calculate the multi-resolution SNRenv.
+        """Calculates the multi-resolution SNRenv.
 
         Parameters
         ----------
@@ -135,32 +158,30 @@ class MrSepsm(Sepsm):
 
         Returns
         -------
-        snr_env_matrix : ndarray
-            Time-average values of the mr-SNRenv.
-        exc_ptns : list of masked_arrays
-            Multi-resolution values of the mixture and of the noise alone.
         mr_snr_env : masked_array
             Multi-resolution `masked_array` of SNRenv.
+        exc_ptns : list of masked_arrays
+            Multi-resolution values of the mixture and of the noise alone.
 
         """
 
-        # noisefloor cannot exceed the mix, since they exist at the same time
+        # First we limit the noise such that it cannot exceed the mix,
+        # since they exist at the same time...
         p_noise = np.minimum(p_noise, p_mix)
 
-        # the noise floor restricted to minimum 0.01 reflecting an
-        # internal noise threshold
+        # ... then we restrict the noise floor restricted to set value
+        # reflecting an internal noise threshold...
         p_mix = np.maximum(p_mix, self.noise_floor)
         p_noise = np.maximum(p_noise, self.noise_floor)
 
-        # calculation of snrenv
+        # ... and we calculate the snrenv according to Eq (2) in Joergensen
+        # et al. (2013)...
         mr_snr_env = (p_mix - p_noise) / p_noise
         mr_snr_env = np.maximum(mr_snr_env, self.snr_env_limit)
         if self.snr_env_ceil is not None:
             mr_snr_env = np.minimum(mr_snr_env, self.snr_env_ceil)
 
-        snr_env_matrix = self._time_average(mr_snr_env)
-
-        return snr_env_matrix, [p_mix, p_noise], mr_snr_env
+        return mr_snr_env, [p_mix, p_noise]
 
     def predict(self, clean, mixture, noise):
         """Predicts intelligibility using the mr-sEPSM.
@@ -184,98 +205,46 @@ class MrSepsm(Sepsm):
 
 
         """
+        signals = np.vstack((clean, mixture, noise))
 
-        fs_new = self.fs / self.downsamp_factor
-        n_clean = len(clean)
-        n_modf = len(self.modf)
-        n_cf = len(self.cf)
+        bands_above_thres_idx = self._find_bands_above_thres(mixture)
+        channel_sigs = self._peripheral_filtering(signals)
+        channel_envs = self._extract_env(channel_sigs)
+        channel_envs = self._mod_sensitivity(channel_envs)
+        filtered_envs, lt_exc_ptns = self._mod_filtering(channel_envs)
+        mr_exc_ptns = self._mr_env_powers(channel_envs, filtered_envs)
+        mr_snr_env_matrix, _ = self._mr_snr_env(
+            mr_exc_ptns[-2], mr_exc_ptns[-1]
+        )
+        time_av_snr_env_matrix = self._time_average(mr_snr_env_matrix)
 
-        # Process only the mixture and noise if the clean speech is the same
-        # as the noise.
-        if (clean is None) or (np.array_equal(clean, mixture)):
-            signals = (mixture, noise)
-        else:
-            signals = (clean, mixture, noise)
-
-        downsamp_chan_envs = np.zeros((len(signals),
-                                       np.ceil(n_clean / self.downsamp_factor)
-                                       .astype('int')))
-        if (downsamp_chan_envs.shape[-1] % 2) == 0:
-            len_offset = 1
-        else:
-            len_offset = 0
-        chan_mod_envs = np.zeros((len(signals),
-                                  len(self.modf),
-                                  downsamp_chan_envs.shape[-1] - len_offset))
-        time_av_mr_snr_env_matrix = np.zeros((n_cf, n_modf))
-        lt_exc_ptns = np.zeros((len(signals), n_cf, n_modf))
-        mr_snr_env_matrix = []
-        mr_exc_ptns = []
-
-        # find bands above threshold
-        filtered_rms_mix = inner.noctave_filtering(mixture, self.cf,
-                                                   self.fs, width=3)
-        bands_above_thres_idx = self._bands_above_thres(filtered_rms_mix)
-
-        for idx_band in bands_above_thres_idx:
-            channel_envs = \
-                [self._peripheral_filtering(signal, self.cf[idx_band])
-                 for signal in signals]
-
-            for i_sig, channel_env in enumerate(channel_envs):
-                # Extract envelope
-                tmp_env = inner.hilbert_envelope(channel_env).squeeze()
-
-                # Low-pass filtering
-                tmp_env = inner.lowpass_env_filtering(tmp_env, 150.0,
-                                                      n=1, fs=self.fs)
-                # Downsample the envelope for faster processing
-                downsamp_chan_envs[i_sig] = tmp_env[::self.downsamp_factor]
-
-                # Sub-band modulation filtering
-                lt_exc_ptns[i_sig, idx_band], chan_mod_envs[i_sig] = \
-                    central.mod_filterbank(downsamp_chan_envs[i_sig],
-                                           fs_new,
-                                           self.modf)
-
-            chan_mr_exc_ptns = []
-            for chan_env, mod_envs in zip(downsamp_chan_envs,
-                                          chan_mod_envs):
-                chan_mr_exc_ptns.append(self._mr_env_powers(chan_env, mod_envs))
-            mr_exc_ptns.append(chan_mr_exc_ptns)
-
-            time_av_mr_snr_env_matrix[idx_band], _, chan_mr_snr_env_matrix \
-                = self._mr_snr_env(*chan_mr_exc_ptns[-2:])  # Select only the
-            # env powers from the mixture and the noise, even if we
-            # calculated the envelope powers for the clean speech.
-            mr_snr_env_matrix.append(chan_mr_snr_env_matrix)
-
-        lt_snr_env_matrix = super(MrSepsm, self)._snr_env(*lt_exc_ptns[-2:])
+        lt_snr_env_matrix, _ = super(MrSepsm, self)._snr_env(*lt_exc_ptns[-2:])
         lt_snr_env = super(MrSepsm, self)._optimal_combination(
-            lt_snr_env_matrix, bands_above_thres_idx)
+            lt_snr_env_matrix, bands_above_thres_idx
+        )
 
         snr_env = self._optimal_combination(
-            time_av_mr_snr_env_matrix, bands_above_thres_idx)
+            time_av_snr_env_matrix, bands_above_thres_idx
+        )
 
         res = {
             'p': {
                 'snr_env': snr_env,
                 'lt_snr_env': lt_snr_env,
             },
-            'snr_env_matrix': time_av_mr_snr_env_matrix,
-            # .snr_env_matrix': snr_env_matrix
+            'snr_env_matrix': time_av_snr_env_matrix,
 
             # Output of what is essentially the sEPSM.
             'lt_snr_env_matrix': lt_snr_env_matrix,
             'lt_exc_ptns': lt_exc_ptns,
 
-            # .mr_snr_env': mr_snr_env
             'mr_snr_env_matrix': mr_snr_env_matrix,
             'mr_exc_ptns': mr_exc_ptns,
 
             'bands_above_thres_idx': bands_above_thres_idx
         }
         return res
+
 
     def _optimal_combination(self, snr_env, bands_above_thres_idx):
         """Combines SNRenv across audio and modulation channels.
@@ -297,11 +266,15 @@ class MrSepsm(Sepsm):
 
         """
         # Acceptable modulation frequencies
-        ma = (np.tile(np.asarray(self.modf), (len(self.cf), 1)).T
-              >= np.asarray(self.cf)).T
+        ma = np.tile(np.asarray(self.modf), (len(self.cf), 1)) \
+             >= np.asarray(self.cf)[:, np.newaxis]
+        # Set the reject SNRenvs to 0. It's ok because we're just doing a
+        # summation
         snr_env[ma] = 0.
+        # Combine across modulation filters...
         snr_env = np.sqrt(np.sum(snr_env[bands_above_thres_idx] ** 2,
                                  axis=-1))
+        # ... and peripheral filters.
         snr_env = np.sqrt(np.sum(snr_env ** 2))
         return snr_env
 
@@ -493,3 +466,4 @@ class MrSepsm(Sepsm):
 
         # fig.text(0.05, 0.5, ylabel, va='center', rotation=90, size=11)
         return fig
+